Unify MultipleLLMPromptExecutorIntegrationTest and SingleLLMPromptExecutorIntegrationTest
